Update virt-install.sh
Redoes implementation of Artix-Runit fix (see issue #2), and removes redundant code for Yum.

+++ virt-install.sh
@@ -1,19 +1,6 @@
 #!/bin/bash
 set -e
 sudo -v
-#fix for Artix Runit systems (see issue #2)
-artix_runit_fix=("sudo ln -s /etc/runit/sv/libvirtd /run/runit/service"
-	"sudo ln -s /etc/runit/sv/virtlogd /run/runit/service"
-	"sudo ln -s /etc/runit/sv/virtlockd /run/runit/service"
-	"sudo sv up libvirtd"
-	"sudo sv up virtlogd"
-	"sudo sv up virtlockd"
-	"sudo sed -i 's/#unix_sock_group = "libvirt"/unix_sock_group = "libvirt"/g' /etc/libvirt/libvirtd.conf"
-	"sudo sed -i 's/#unix_sock_ro_perms = "0777"/unix_sock_ro_perms = "0777"/g' /etc/libvirt/libvirtd.conf"
-	"sudo sed -i 's/#unix_sock_rw_perms = "0770"/unix_sock_rw_perms = "0770"/g' /etc/libvirt/libvirtd.conf"
-	"sudo usermod -aG libvirt $USER"
-	"echo "Installation complete, restart your system for changes to take effect.""
-)
 #Downloads KVM, QEMU, Virt-manager, all its dependencies, and a few other tools.
 echo "Downloading dependencies"
 if command -v pacman >/dev/null; then #Arch
@@ -22,12 +9,9 @@ if command -v pacman >/dev/null; then #Arch
 elif command -v apt-get >/dev/null; then #Debian/Ubuntu
 	sudo apt-get update
 	sudo apt-get install -y qemu-kvm libvirt-daemon-system libvirt-clients bridge-utils virtinst virt-manager
-elif command -v dnf >/dev/null; then #Fedora
+elif command -v dnf >/dev/null; then #Fedora, Red Hat, & CentOS
 	sudo dnf update
 	sudo dnf install -y qemu-kvm libvirt libvirt-client virt-install virt-manager
-elif command -v yum >/dev/null; then #Red Hat/Cent OS
-	sudo yum update
-	sudo yum install -y qemu-kvm libvirt libvirt-client virt-install virt-manager
 elif command -v emerge >/dev/null; then #Gentoo
 	echo "WARNING: Untested Distro!"
 	sudo emerge --sync
@@ -48,10 +32,6 @@ fi
 if [[ -f "/etc/artix-release" ]]; then #Artix
 if command -v rc-status >/dev/null; then #OpenRC
 	sudo pacman -S libvirt-openrc
-elif command -v sv >/dev/null; then #Runit
-	sudo pacman -S libvirt-runit
-	eval "${artix_runit_fix[@]}"
-	exit 0
 elif command -v dinitctl >/dev/null; then #Dinit
 	sudo pacman -S libvirt-dinit
 elif command -v s6-rc >/dev/null; then #s6
@@ -69,8 +49,17 @@ elif command -v rc-status >/dev/null; then #OpenRC
 	sudo rc-update add libvirtd
 	sudo rc-service libvirtd start
 elif command -v sv >/dev/null; then #Runit
-	sudo ln -s /etc/sv/libvirtd /etc/runit/runsvdir/default/
-	sudo sv up libvirtd
+	if [[ -f "/etc/artix-release" ]]; then #Artix-Runit
+		sudo ln -s /etc/runit/sv/libvirtd /run/runit/service
+		sudo ln -s /etc/runit/sv/virtlogd /run/runit/service
+		sudo ln -s /etc/runit/sv/virtlockd /run/runit/service
+		sudo sv up libvirtd
+		sudo sv up virtlogd
+		sudo sv up virtlockd
+	else
+		sudo ln -s /etc/sv/libvirtd /etc/runit/runsvdir/default/
+		sudo sv up libvirtd
+	fi #This whole section is untested and is intended to fix issue #2
 elif command -v dinitctl >/dev/null; then #Dinit
 	sudo dinitctl start libvirtd
 	sudo dinitctl enable libvirtd
@@ -91,9 +80,6 @@ sudo sed -i 's/#unix_sock_group = "libvirt"/unix_sock_group = "libvirt"/g' /etc/
 sudo sed -i 's/#unix_sock_ro_perms = "0777"/unix_sock_ro_perms = "0777"/g' /etc/libvirt/libvirtd.conf
 sudo sed -i 's/#unix_sock_rw_perms = "0770"/unix_sock_rw_perms = "0770"/g' /etc/libvirt/libvirtd.conf
 #comment out the above 3 lines and uncomment the 3 lines below in case it ends up not working
-#sudo awk '/^#.*unix_sock_group/{sub(/^#/,"",$0)}1' /etc/libvirt/libvirtd.conf > temp && sudo mv temp /etc/libvirt/libvirtd.conf
-#sudo awk '/^#.*unix_sock_ro_perms/{sub(/^#/,"",$0)}1' /etc/libvirt/libvirtd.conf > temp && sudo mv temp /etc/libvirt/libvirtd.conf
-#sudo awk '/^#.*unix_sock_rw_perms/{sub(/^#/,"",$0)}1' /etc/libvirt/libvirtd.conf > temp && sudo mv temp /etc/libvirt/libvirtd.conf
 # Adds current user to the libvirt group
 sudo usermod -aG libvirt $USER
 echo "Installation complete, restart your system for changes to take effect."
